# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Quin Travel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/news/list_news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get list news",
                "description": "Returns list news information",
                "operationId": "list_news",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page of news to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit of news to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                }
            }
        },
        "/api/news/{slug}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get News Information",
                "description": "Returns news information",
                "operationId": "get news detail",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "News slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "description": "JWT Token (Bearer <token>)"
            }
        }
    }
}